Upgrade the action output code to use env files
Using the `::set-output` has been suddenly deprecated the other day,
without a clear deadline [1]. The new way of achieving the same is to
append key-value pairs to the environment file that can be located by
looking up the path recorded in the `$GITHUB_OUTPUT` environment
variable [2].

This patch implements this new mechanism. The change is not a
user-facing update and the behavior should remain the same.

Resolves https://github.com/re-actors/alls-green/issues/11.

[1]: https://hynek.me/til/set-output-deprecation-github-actions/
[2]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter